src/test/java/com/example/demo/service/AnimalServiceTest.java
package com.example.demo.service;

import com.example.demo.domain.animal.Animal;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.junit.jupiter.api.extension.ExtendWith;
import org.mockito.InjectMocks;
import org.mockito.Mock;
import org.mockito.junit.jupiter.MockitoExtension;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.beans.factory.annotation.Qualifier;
import org.springframework.boot.test.context.SpringBootTest;

import static org.junit.jupiter.api.Assertions.*;
import static org.mockito.Mockito.doNothing;
import static org.mockito.Mockito.verify;
import static org.mockito.internal.verification.VerificationModeFactory.times;

@SpringBootTest
class AnimalServiceTest {
    @Autowired
    private AnimalService animalService;

    @Test
    void testAnimalSound() {
        animalService.animalSound();
        //Assertions.assertThat(sound).isEqualTo("Bark");
    }

}